Fixes #1
Changed the implementation of classes referenced by the
ArtifactFileUploader to be Serializable. Updated the uploader's
implementation as necessary. Removed reference to abstract build, which
is not serializable, pass the build's result directly.

This fixes the issue of crashing with a serialization exception in a
master/slave setup, due to the job not being fully serializable. The
file uploader can now be correctly serialized, artifacts built on a
slave node are now uploaded.