Show the actual value of a default option in a publication's settings.
If one of a publication's settings is set to "default" it will now show
the name of the option that has been configured in the global settings.
This makes is easier to see what the default value is.

Before this change the setting would show "(default)" as the selected
item. This required the user to open the global settings to see the
configured default value. The setting will now show "name* (default)",
where name is the name of the configured option.