Add the supported DH key pair size in bits to the debug log output.
This adds log entries that show whether a key pair size of 1024, 2048
and 4096 bits for DiffieHellman is supported by the current platform and
Java VM. For example, if a key pair size of 2048 bits is unsupported and
the store's SSL certificate uses a key with 2048 bits the upload will
fail. See issue #3.