Upload the change log and description for each language of a server.
Fixes #2. The artifact uploader will now query the languages configured
on the server to determine for which languages the change log and
description need to be uploaded. This fixes the issue of not actually
setting the change log or description of a version if previous versions
had no text set (manually).