Replace entities with plain JSON objects and arrays.
This replaces all Java objects that mirror the server's entities with
plain JSON (Gson's JsonObject/JsonArray). This fixes the issue of having
to keep entities in sync with the server.

E.g. if a new field was added to one of the server's entities, this
field would get lost during the conversion, unless the plugin was
updated to add the same field to its entity-class. Since the plugin now
uses plain JSON objects this is no longer the case.

This should make the plugin more resilient to server side changes.