Add Android test app that can be built on Jenkins
This adds a very simple Android application project that can be built on a Jenkins server and uploaded to a Relution server using the plugin. This allows to test the plugin without the need to build an actual full-size Android project.

Additionally artifacts are versioned using the build number, so each artifact represents a new version.